missing & caused a corrupt memory bug that never caused a crash. my god.
