+++ T3/inputs/skim_deepgen_tmpl.py
@@ -39,7 +39,7 @@ def fn(input_name, isData, full_path):
         processType = root.kTop
     analysis = deepgen() 
     analysis.processType=processType 
-    analysis.deepGenGrid = True
+#    analysis.deepAntiKtSort = True
     analysis.dump()
     skimmer.SetAnalysis(analysis)
     skimmer.isData=isData
