+++ data/deep/flatten_gen.root
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:55ab9022c15dc8b09b832b5fa011444d7ef51da8f0748885044e160c6bd70af8
-size 7349
+oid sha256:b8cf4365a618948ceb333e1d7045d5b2d43483f7c2845aaf0f8c872d7cf86729
+size 7352
