+++ data/deep/flatten_gen_scaled.root
@@ -1,3 +1,3 @@
 version https://git-lfs.github.com/spec/v1
-oid sha256:da54db24071e31fae57eb7ce7090f98d1fba8648a04a390ffe8698db35e103bf
+oid sha256:61ae1fcb8b53220a072237b03dea548ace210155e1fd9a6994c89d21a29f5535
 size 7356
